// src/app/module/auth/auth.interface.ts
import {
	Role,
	Gender,
} from "../../../../prisma/generated/prisma/enums";

export interface ILoginUserPayload {
	email: string;
	password: string;
}

export interface IRegisterClientPayload {
	name: string;
	email: string;
	password: string;

	client?: {
		contactNumber?: string;
		address?: string;
		gender?: Gender;
		dateOfBirth?: string;
	};
}

export interface IVerifyEmailPayload {
	email: string;
	otp: string;
}

export interface IRequestUser {
	userId: string;
	email: string;
	name: string;
	role: Role;
}

export interface IGoogleLoginPayload {
	idToken: string;
}

export interface IForgotPasswordPayload {
	email: string;
}

export interface IResetPasswordPayload {
	email: string;
	newPassword: string;
	otp: string;
}